PDF_Chatbot/requirements.txt
# -- GEREKLİ TÜM KÜTÜPHANELER --
streamlit
google-generativeai
langchain
langchain-community
langchain-google-genai
faiss-cpu
PyMuPDF
python-dotenv